Rename README.md to Phase Equilibrium in Aqueous Two-phase Systems Containing Poly (propylene glycol) and Sodium Citrate at Different pH
Biphasic systems formed by mixing of two polymers or a polymer and a salt in water can be used for separation of cells, membranes, viruses, proteins, nucleic acids, and other biomolecules. The partitioning between the two phases is dependent on the surface properties and conformation of the materials, and also on the composition of the two-phase system. The mechanism of partitioning is, however, complex and not easily predicted. Aqueous two-phase systems (ATPS) have proven to be a useful tool for analysis of biomolecular and cellular surfaces and their interactions, fractionation of cell populations, product recovery in biotechnology, and so forth. Potential for environmental remediation has also been suggested. Because ATPS are easily scalable and are also able to hold high biomass load in comparison with other separation techniques, the application that has attracted most interest so far has been the large-scale recovery of proteins from crude feedstocks. As chemicals constitute the major cost factor for large-scale systems, use of easily recyclable phase components and the phase systems generated by a single-phase chemical in water are being studied.